Refactoring (166-faucet)
