+++ evm_loader/faucet/src/main.rs
@@ -4,7 +4,6 @@
 #![forbid(unsafe_code)]
 #![deny(warnings)]
 
-mod airdrop;
 mod cli;
 mod config;
 mod ethereum;
